+++ src/dfx/src/config/cache.rs
@@ -131,6 +131,7 @@ pub fn install_version(v: &str, force: bool) -> Result<PathBuf, CacheError> {
 
         // Copy our own binary in the cache.
         let dfx = temp_p.join("dfx");
+        #[allow(clippy::needless_borrow)]
         dfx_core::fs::write(
             &dfx,
             dfx_core::fs::read(&current_exe).map_err(UnifiedIoError::from)?,